Refactor weather service and improve API configurations
Removed unused countryCode parameter from GetCoordinatesFromCityAsync. Added error codes and validation for latitude/longitude in WeatherRepository. Refactored API setup for improved clarity, updated Swagger descriptions, and switched data types for consistency in Weather-related DTOs and entities.

+++ DTO/Output/GetWeatherByLocationOutputDto.cs
@@ -3,7 +3,7 @@
 public class GetWeatherByLocationOutputDto
 {
     public double Temperature { get; init; }
-    public int WindDirection { get; init; }
-    public int WindSpeed { get; init; }
+    public double WindDirection { get; init; }
+    public double WindSpeed { get; init; }
     public DateTime SunriseDateTimeIso8601 { get; init; }
 }
