+++ Entities/Weather.cs
@@ -12,7 +12,7 @@ public class Weather
     public double Longitude { get; set; }
     public double Latitude { get; set; }
 
-    public int WindDirection { get; set; }
+    public double WindDirection { get; set; }
     public double Temperature { get; set; }
     public double WindSpeed { get; set; }
 
@@ -40,7 +40,7 @@ public class OpenMeteoResponse
 public class CurrentWeather
 {
     [JsonPropertyName("wind_direction_10m")]
-    public int WindDirection { get; set; }
+    public double WindDirection { get; set; }
 
     [JsonPropertyName("temperature_2m")]
     public double Temperature { get; set; }
